Implemented extensive AI incident, drift, and SLA governance enhancements:
- Added AI Incident Response Procedure with full incident classification, lifecycle, and escalation, plus ISO/NIST alignment
- Introduced AI Incident fields, incident types, and reporting workflows
- Created AI Model Drift Monitoring with drift indicators, thresholds, and response protocols
- Implemented AI SLA configurations for human review (AISLAHumanReviewConfig) and SLA dashboards
- Added incident and drift related documentation notes and UI components integration
- Extended coaching prompts UI to include field references and database architecture notes
- Updated sentiments and writing quality analysis to reflect new data architectures and fields
- Exported new components and updated index for broader usage

X-Lovable-Edit-ID: edt-e3eb64ff-5506-44b5-a244-5b47c83ac10f

+++ src/components/enablement/manual/feedback360/sections/ai/AIWritingQualityAnalysis.tsx
@@ -17,7 +17,7 @@ const qualityFields: FieldDefinition[] = [
   {
     name: 'clarity_score',
     required: false,
-    type: 'integer',
+    type: 'decimal',
     description: 'How clear and understandable the feedback is (0-100)',
     defaultValue: 'null',
     validation: '0-100'
@@ -25,7 +25,7 @@ const qualityFields: FieldDefinition[] = [
   {
     name: 'specificity_score',
     required: false,
-    type: 'integer',
+    type: 'decimal',
     description: 'How specific vs. vague the feedback is (0-100)',
     defaultValue: 'null',
     validation: '0-100'
@@ -33,7 +33,7 @@ const qualityFields: FieldDefinition[] = [
   {
     name: 'bias_risk_score',
     required: false,
-    type: 'integer',
+    type: 'decimal',
     description: 'Inverse of bias detection - higher is less biased (0-100)',
     defaultValue: 'null',
     validation: '0-100'
@@ -41,7 +41,7 @@ const qualityFields: FieldDefinition[] = [
   {
     name: 'behavioral_focus_score',
     required: false,
-    type: 'integer',
+    type: 'decimal',
     description: 'How focused on observable behaviors vs. personality (0-100)',
     defaultValue: 'null',
     validation: '0-100'
@@ -49,7 +49,7 @@ const qualityFields: FieldDefinition[] = [
   {
     name: 'overall_quality_score',
     required: false,
-    type: 'integer',
+    type: 'decimal',
     description: 'Weighted composite of all dimensions (0-100)',
     defaultValue: 'null',
     validation: '0-100'
